Consolidate writing style functions
