+++ src/lib/writingStyle.ts
@@ -1,4 +1,6 @@
 import { DraftBeaconUse } from "../entities/DraftBeaconUse";
+import { Use } from "../entities/Use";
+import { Activity } from "./deprecatedRegistration/types";
 
 export const prettyUseName = (use: DraftBeaconUse): string =>
   makeEnumValueUserFriendly(use.environment) +
@@ -88,4 +90,27 @@ export const formatMonth = (dateString: string): string => {
     month: "long",
     year: "numeric",
   });
+};
+
+export const formatUses = (uses: Use[]): string =>
+  uses.reduce((formattedUses, use, index, uses) => {
+    if (index === uses.length - 1) return formattedUses + formatUse(use);
+    return formattedUses + formatUse(use) + ", ";
+  }, "");
+
+export const formatUse = (use: Use): string => {
+  const formattedActivity =
+    use.activity === Activity.OTHER
+      ? titleCase(use.otherActivity || "")
+      : titleCase(use.activity);
+  const formattedPurpose = use.purpose ? ` (${titleCase(use.purpose)})` : "";
+  return formattedActivity + formattedPurpose;
+};
+
+export const titleCase = (text: string): string => {
+  return text
+    .replace(/_/g, " ")
+    .split(" ")
+    .map((word) => (word[0] || "").toUpperCase() + word.slice(1).toLowerCase())
+    .join(" ");
 };
