--- src/utils/formatUses.ts
@@ -1,25 +1,0 @@
-import { Use } from "../entities/Use";
-import { Activity } from "../lib/deprecatedRegistration/types";
-
-export const formatUses = (uses: Use[]): string =>
-  uses.reduce((formattedUses, use, index, uses) => {
-    if (index === uses.length - 1) return formattedUses + formatUse(use);
-    return formattedUses + formatUse(use) + ", ";
-  }, "");
-
-export const formatUse = (use: Use): string => {
-  const formattedActivity =
-    use.activity === Activity.OTHER
-      ? titleCase(use.otherActivity || "")
-      : titleCase(use.activity);
-  const formattedPurpose = use.purpose ? ` (${titleCase(use.purpose)})` : "";
-  return formattedActivity + formattedPurpose;
-};
-
-export const titleCase = (text: string): string => {
-  return text
-    .replace(/_/g, " ")
-    .split(" ")
-    .map((word) => (word[0] || "").toUpperCase() + word.slice(1).toLowerCase())
-    .join(" ");
-};
